Move PluginRegistry in coreutils
