+++ packages/coreutils/package.json
@@ -41,6 +41,9 @@
     "test:webkit-headless": "cd tests && karma start --browsers=WebkitHeadless",
     "watch": "tsc --build --watch"
   },
+  "dependencies": {
+    "@lumino/algorithm": "^2.0.1"
+  },
   "devDependencies": {
     "@lumino/buildutils": "^2.0.1",
     "@microsoft/api-extractor": "^7.36.0",
