+++ review/api/coreutils.api.md
@@ -5,6 +5,18 @@
 ```ts
 
 // @public
+export interface IPlugin<T, U> {
+    activate: (app: T, ...args: any[]) => U | Promise<U>;
+    autoStart?: boolean | 'defer';
+    deactivate?: ((app: T, ...args: any[]) => void | Promise<void>) | null;
+    description?: string;
+    id: string;
+    optional?: Token<any>[];
+    provides?: Token<U> | null;
+    requires?: Token<any>[];
+}
+
+// @public
 export interface JSONArray extends Array<JSONValue> {
 }
 
@@ -65,6 +77,37 @@ export interface PartialJSONObject {
 
 // @public
 export type PartialJSONValue = JSONPrimitive | PartialJSONObject | PartialJSONArray;
+
+// @public
+export class PluginRegistry<T = any> {
+    constructor(options?: PluginRegistry.IOptions);
+    activatePlugin(id: string): Promise<void>;
+    activatePlugins(kind: 'startUp' | 'defer', options?: PluginRegistry.IStartOptions): Promise<void>;
+    get application(): T;
+    set application(v: any);
+    deactivatePlugin(id: string): Promise<string[]>;
+    get deferredPlugins(): string[];
+    deregisterPlugin(id: string, force?: boolean): void;
+    getPluginDescription(id: string): string;
+    hasPlugin(id: string): boolean;
+    isPluginActivated(id: string): boolean;
+    listPlugins(): string[];
+    registerPlugin(plugin: IPlugin<T, any>): void;
+    registerPlugins(plugins: IPlugin<T, any>[]): void;
+    resolveOptionalService<U>(token: Token<U>): Promise<U | null>;
+    resolveRequiredService<U>(token: Token<U>): Promise<U>;
+}
+
+// @public
+export namespace PluginRegistry {
+    export interface IOptions {
+        validatePlugin?: (plugin: IPlugin<any, any>) => boolean;
+    }
+    export interface IStartOptions {
+        ignorePlugins?: string[];
+        startPlugins?: string[];
+    }
+}
 
 // @public
 export class PromiseDelegate<T> {
